Stylize basic HTML website with CSS
The basic recipe website with the homepage and three recipes made using HTML was styled using CSS. This involved the use of selectors and various changes regarding colors, fonts, and layout to improve the website's appearance and also the usability. External CSS was used in the form of styles.css for the 3 recipe pages and one for the homepage.